+++ demos/manual/src/main/java/is/codion/framework/demos/manual/common/InputControls.java
@@ -201,16 +201,28 @@ public final class InputControls {
     // end::booleanComboBox[]
   }
 
-  static void textField() {
-    // tag::textField[]
+  static void stringField() {
+    // tag::stringField[]
     Value<String> stringValue = Value.value();
 
     JTextField textField =
-            Components.textField(stringValue)
+            Components.stringField(stringValue)
                     .preferredWidth(120)
                     .transferFocusOnEnter(true)
                     .build();
-    // end::textField[]
+    // end::stringField[]
+  }
+
+  static void characterField() {
+    // tag::characterField[]
+    Value<Character> characterValue = Value.value();
+
+    JTextField textField =
+            Components.characterField(characterValue)
+                    .preferredWidth(120)
+                    .transferFocusOnEnter(true)
+                    .build();
+    // end::characterField[]
   }
 
   static void textArea() {