+++ swing/common-ui/src/test/java/is/codion/swing/common/ui/component/ComponentsTest.java
@@ -576,9 +576,9 @@ public final class ComponentsTest {
   }
 
   @Test
-  void textField() {
+  void stringField() {
     Value<String> value = Value.value();
-    ComponentValue<String, JTextField> componentValue = Components.textField()
+    ComponentValue<String, JTextField> componentValue = Components.stringField()
             .columns(10)
             .upperCase(true)
             .selectAllOnFocusGained(true)
@@ -796,9 +796,9 @@ public final class ComponentsTest {
 
   @Test
   void valueLocked() {
-    assertThrows(IllegalStateException.class, () -> Components.textField(Value.value())
+    assertThrows(IllegalStateException.class, () -> Components.stringField(Value.value())
             .linkedValue(Value.value()));
-    assertThrows(IllegalStateException.class, () -> Components.textField(Value.value())
+    assertThrows(IllegalStateException.class, () -> Components.stringField(Value.value())
             .linkedValue(Value.value()));
   }